requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==1.4.4
Brotli==1.1.0
certifi==2019.3.9
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.1
cssselect2==0.7.0
defusedxml==0.6.0
Django==5.1.6
django-allauth==65.4.1
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
fonttools==4.56.0
html5lib==1.1
idna==2.8
lxml==5.3.1
oauthlib==3.0.1
oscrypto==1.3.0
pep8==1.7.1
pillow==11.1.0
pycodestyle==2.5.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.3.0
pyphen==0.17.2
python-bidi==0.6.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==1.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
